Handle changed Keywords.keys/1 interface
The interface for Keyword.keys/1 changed slightly in Elixir v1.11.0 and
now checks its input for a valid list of 2 length tuples. As the
current implementation can generate lists with tuples which look like

	[{{:payments, :id}, :asc}]

or

	[{:payments, :id}]

we need to normalise them under the hood. This new private function does
just that.